Pets no longer teleport on unload when sitting
# Changes
* Pets will no longer teleport to the player when they are sitting and the chunk they are in unloads.

# Reason for change
Randomly, all your pets will teleport to you when running in the nether - even if sitting. This happens when the previous chunk unloaded and so now all your lovely pets are now on fire, burning, dying, and may about to try swimming in lava. I've lost pets before to this, and I'm sick of roaming the nether without being able to leave my pets behind. This should fix that change.

Thanks for reading. :)